requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-core>=0.1.0
langchain-community>=0.0.10
openai>=1.0.0
tavily-python>=0.3.1
beautifulsoup4>=4.12.0
requests>=2.28.0
streamlit>=1.27.0
python-dotenv>=1.0.0
firecrawl>=0.1.0
fpdf2>=2.7.5
markdown>=3.5.1
playwright>=1.40.0
html2text>=2020.1.16
pymupdf>=1.23.0
unstructured>=0.11.0
unstructured-inference>=0.7.0
numpy>=1.24.0
scikit-learn>=1.3.0
# Airtable integration
airtable-python-wrapper>=0.15.0 